layout/base/tests/test_bug559499.html

<!DOCTYPE HTML>
<html style="background:yellow">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=559499
-->
<head>
  <title>Test for Bug 559499</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body style="position:relative; z-index:-1; padding-top:100px;">
<p><a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=559499">Mozilla Bug 514127</a></p>

<pre id="test">
<script type="application/javascript">

/** Test for Bug 559499 **/

is(document.elementFromPoint(50, 50), document.body, "Able to hit body");
document.documentElement.style.display = "table";
is(document.elementFromPoint(50, 50), document.body, "Able to hit body (table)");

</script>
</pre>
</body>
</html>
